Add table of contents and remove comment on extension from restore
In order to support restores where we don't have superuser permissions
we must not restore the comment on the TimescaleDB extension, because
comments can only be modified by owners of objects or superusers,
extensions have no owners, therefore, you must be a superuser to modify
the comment on an extension, as far as I can tell.

This also introduces the infrastructure necessary to filter further
output from commands, so we will be able to use it in the future to
filter out lines containing warnings or other messages that can safely
be ignored by our users.