Adds a pg_dumpall step to dump roles/tablespaces
Roles and tablespaces are globals that are not dumped with a typical
pg_dump of a single database. Forgetting to dump the roles and
tablespaces on a cluster means that a dumped database can be quite
difficult to restore as there are, by default grants etc. included
in a dump. There is the possibility of removing them from the restore
but it seems better to include the dump of roles.

Restoring them is more annoying as it can only happen once, because of
that, I haven't included them in the restore process, but rather just
included the file, which can be run with a command like
`psql -d [db-URI] -f [dump-dir]/roles.sql`