Allow all pg_dump/pg_restore parameters
Before this commit there was no way to influence the
`pg_dump`/`pg_restore` behavior directly by specifying certain flags.
Some issues however, (like privileges, ownership) were unable to passed
along to the `pg_restore` command.

By passing along all parameters after the double dash -- to the binary
that is being called, we open up the possibility of using the full
feature set of the `pg_dump` and `pg_restore` tools.

The downside of adding this is that some flags may not play nicely with
the `timescaledb-backup` native flags themselves. Those issues will need
to be documented/taken care of when they arise.